[Multicol] Start adding performance tests for the multi-column implementation https://bugs.webkit.org/show_bug.cgi?id=137687
Reviewed by Mihnea Ovidenie.

This patch creates a Multicol folder inside the Layout performance tests suite. It adds only two tests
as described below.

The tests are skipped until the implementation stabilizes.

* Layout/Multicol/MulticolManyColumns.html: Added.
This test verifies the performance of the multi-column implementation with two nested multi-column
containers, the first having 20 columns and the second 10 columns.

* Layout/Multicol/MulticolNested.html: Added.
This test verifies the performance of nesting multi-column containers in the presence
of floats.

* Layout/Multicol/resources/multicol-content-many-columns.html: Added.
* Layout/Multicol/resources/multicol-content-nested.html: Added.
* Skipped: Skip the Multicol folder for now.


Identifier: 155587@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@174726 268f45cc-cd09-0410-ab3c-d52691b4dbfc