Display legal text when uploading a patch with a new bug
* js/attachment.js:
(setContentTypeDisabledState): Make this work on the Enter Bug
page so that the "Submit Bug" button changes to
"Agree and Submit Bug" when the "patch" checkbox is checked, and
it changes back to "Submit Bug" when unchecked.
* template/en/custom/attachment/createformcontents.html.tmpl:
Include the attachment/patchlegalese.html.tmpl template when
returning the page so that it will be shown when the "patch"
checkbox is checked.  Set the 'product' variable from the CGI
variable of the same name.
* template/en/custom/attachment/patchlegalese.html.tmpl: Add the
'expert_fields' class to the <tr> element so that it gets
shown/hidden when the Show/Hide Advanced Fields links are
clicked.

Identifier: 155631@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@174772 268f45cc-cd09-0410-ab3c-d52691b4dbfc