Remove references to Rietveld from bugs.webkit.org
Rietveld was an experiment that was added by Google.  It was
never enabled for everyone, so it's safe to remove the changes
that reference it.

* Bugzilla/Flag.pm: Remove code that prevented mail messages from
being sent for patches reviewed in Rietveld.
* attachment.cgi: Remove reference to the "rietveldreview"
template.
* template/en/custom/attachment/list.html.tmpl: Remove code
that handled the Rietveld review flag.
* template/en/custom/attachment/reviewform.html.tmpl: Remove code
that handled the "rietveldReview" CGI parameter.
* template/en/custom/attachment/rietveldreview.html.tmpl: Delete.
* template/en/custom/flag/list.html.tmpl: Delete.  After
removing the code to handle Rietveld reviews, there were no more
local WebKit changes to the custom template file, so it could be
removed.

Identifier: 155632@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@174773 268f45cc-cd09-0410-ab3c-d52691b4dbfc