[iOS] Fix incorrect interface orientation that can be caused by fullscreen video dismissal. https://bugs.webkit.org/show_bug.cgi?id=137930.
Patch by Reza Abbasian <rabbasian@apple.com> on 2014-11-04
Reviewed by Jer Noble.

When a fullscreen video gets dismissed programmatically, for instance for removing
a webview from the view hierarchy, it can put the device in an incorrect orientation.
Make sure we retain the window of parentView when we attempt to go to fullscreen and use
the retained window for putting the device in correct orientation after dismissal.

* platform/ios/WebVideoFullscreenInterfaceAVKit.h:
* platform/ios/WebVideoFullscreenInterfaceAVKit.mm:
(WebVideoFullscreenInterfaceAVKit::setupFullscreen):
(WebVideoFullscreenInterfaceAVKit::cleanupFullscreen):
(WebVideoFullscreenInterfaceAVKit::invalidate):

Identifier: 156208@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@175585 268f45cc-cd09-0410-ab3c-d52691b4dbfc